# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Platform API Documentation",
        "description": "API documentation for mobile app, vendor integrations, and admin services",
        "contact": {
            "email": "admin@ecommerce-platform.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server (localhost)"
        }
    ],
    "paths": {
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List orders",
                "description": "Returns all orders with customer info and status.",
                "operationId": "6a829a2750ceacdbe340f4fc8782dbd5",
                "responses": {
                    "200": {
                        "description": "Orders list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get order",
                "description": "Returns order details by id.",
                "operationId": "6b3e9c27d4bc04bb3a97dcc31f3f21f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List products",
                "description": "Returns all products with vendor and status information.",
                "operationId": "2ad65e2d8ebe50ffb9403be255e10888",
                "responses": {
                    "200": {
                        "description": "Products list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get product",
                "description": "Returns product details by id.",
                "operationId": "418a8a8b609bfb6de33a8bb7a300fb57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}/approve": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve product",
                "description": "Marks a product as approved.",
                "operationId": "70c54c38cfea8f2fbc69e111a56ea8a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product approved"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}/reject": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Reject product",
                "description": "Marks a product as rejected.",
                "operationId": "e0ab1d414ff05179d9692d14459763b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product rejected"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/admin/vendors": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List vendors",
                "description": "Returns all vendor accounts and approval status.",
                "operationId": "a55f1a0c0261519bef834d5748f51d98",
                "responses": {
                    "200": {
                        "description": "Vendors list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/vendors/{id}/approve": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve a vendor",
                "description": "Sets vendor approval status to true.",
                "operationId": "09b8941fb31de0e2507ab78fccc4b5ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor approved"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/vendors/{id}/block": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Block a vendor",
                "description": "Sets vendor approval status to false.",
                "operationId": "2ec411550d53dcf203f832cb7782dfcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor blocked"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}